% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_domains.R
\name{detect.domains}
\alias{detect.domains}
\title{Check which domains (ATC, EBM, ICD, OPS) occur in each VAR_NAME}
\usage{
detect.domains(
  code.frame,
  grp.vars = c("VAR_GRP", "VAR_NAME"),
  domain.var = "domain"
)
}
\arguments{
\item{code.frame}{data.frame containing the relevant code groups and a
column indicating which domains are to be used (can simply be the
concatenated codelists)}

\item{grp.vars}{Vector of column names for grouping of variables (defaults
to c("VAR_GRP", "VAR_NAME"))}

\item{domain.var}{Character string indicating column name of column giving
domain information (defaults to "domain")}
}
\value{
data.frame
}
\description{
Expects concatenated codelists (one data.frame) resulting from the function
output of read.codelist()
}
