% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logreader.R
\name{logreader}
\alias{logreader}
\title{Scan a Log file for errors, warnings and suspicious messages}
\usage{
logreader(logfile = NULL, r.sas = "SAS")
}
\arguments{
\item{logfile}{Path to the logfile, including file ending}

\item{r.sas}{Either "R" or "SAS", defaults to "SAS"}
}
\value{
Findings as a data.frame
}
\description{
Load a log file and scan it for errors, warnings and suspicious messages.
The function has two modi: 1) if r.sas = R, it expects an R log (e.g. the .Rout
file that is generated when using R from the command line) or 2) if r.sas = SAS, it
expects a .log file from a SAS session.
}
