% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MalteColors.R
\name{MalteColors}
\alias{MalteColors}
\title{Function to return a color palette of length n}
\usage{
MalteColors(n = NULL)
}
\arguments{
\item{n}{Length of desired color vector}
}
\value{
Character vector
}
\description{
Returns a character vector of length n containing Hex values for a color palette
}
\examples{
MalteColors(4)
}
