% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_header.R
\name{new.header}
\alias{new.header}
\title{Print a program skeleton including header to the console}
\usage{
new.header(proj = NULL)
}
\arguments{
\item{proj}{Name of the current project}
}
\value{
Message to the console
}
\description{
Run this function without any arguments and it will print a template for a
new program header to the console. Copy-paste this into the new program and
fill in the relevant information. Version and date are filled in automatically.
}
\examples{
new.header()

}
