% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oracle_checkname.R
\name{oracle.checkname}
\alias{oracle.checkname}
\title{Check variable names for Oracle SQL reserved terms}
\usage{
oracle.checkname(var.name)
}
\arguments{
\item{var.name}{Character vector with variable names to be checked. Not case sensitive.}
}
\value{
Overview as matrix
}
\description{
This function checks, if new variable names are Oracle SQL reserved terms. Variable names
are checked against this list:
https://docs.oracle.com/cd/B19306_01/em.102/b40103/app_oracle_reserved_words.htm
}
\examples{
oracle.checkname(c("ADD", "W_ALCABU"))
}
