% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oracle_contents.R
\name{oracle.contents}
\alias{oracle.contents}
\title{Send Oracle SQL code to collect column names of all variables in user-specified tables}
\usage{
oracle.contents(
  schema,
  tables,
  username = NULL,
  password = NULL,
  dbname = NULL
)
}
\arguments{
\item{schema}{Oracle schema where the tables are stored}

\item{tables}{Character vector containing the names of the tables to be queried}

\item{username}{Oracle username, defaults to NULL.}

\item{password}{Oracle password, defaults to NULL.}

\item{dbname}{Oracle database name, defaults to NULL.}
}
\value{
list object with one element per table containing information on schema, table name, columns, data type and length of variable
}
\description{
This function assumes the following:
\enumerate{
\item The ROracle package is installed and loaded. It must be loaded manually,
because I don't want to include it as dependency, because it needs to
be set up with administrator rights.
\item There is a SAS autoexec file in the usual form which contains
the Oracle user credentials in the usual form. Alternatively,
the credentials can be specified manually in the arguments username,
password and dbname
}
}
