% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oracle_query.R
\name{oracle.query}
\alias{oracle.query}
\title{Send Oracle SQL code to the database and retrieve data}
\usage{
oracle.query(ora.code, username = NULL, password = NULL, dbname = NULL)
}
\arguments{
\item{ora.code}{Oracle SQL code to be sent to the Oracle server}

\item{username}{Oracle username, defaults to NULL.}

\item{password}{Oracle password, defaults to NULL.}

\item{dbname}{Oracle database name, defaults to NULL.}
}
\value{
Data retrieved from Oracle server, as data.frame
}
\description{
This function assumes the following:
\enumerate{
\item The ROracle package is installed and loaded. It must be loaded manually,
because I don't want to include it as dependency, because it needs to
be set up with administrator rights.
\item There is a SAS autoexec file in the usual form which contains
the Oracle user credentials in the usual form. Alternatively,
the credentials can be specified manually in the arguments username,
password and dbname
}
}
