% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oracle_tables.R
\name{oracle.tables}
\alias{oracle.tables}
\title{Send Oracle SQL code to collect names of all tables that are stored
on an Oracle schema}
\usage{
oracle.tables(
  schema,
  username = NULL,
  password = NULL,
  dbname = NULL,
  only_T = FALSE
)
}
\arguments{
\item{schema}{Oracle schema to check for tables}

\item{username}{Oracle username, defaults to NULL.}

\item{password}{Oracle password, defaults to NULL.}

\item{dbname}{Oracle database name, defaults to NULL.}

\item{only_T}{Logical, only query tables starting with "T_", defaults to FALSE. This argument makes sense when querying snapshots, but not when querying project schemas.}
}
\value{
data.frame with names of each table on the schema
}
\description{
This function assumes the following:
\enumerate{
\item The ROracle package is installed and loaded. It must be loaded manually,
because I don't want to include it as dependency, because it needs to
be set up with administrator rights.
\item There is a SAS autoexec file in the usual form which contains
the Oracle user credentials in the usual form. Alternatively,
the credentials can be specified manually in the arguments username,
password and dbname
}
}
