% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_codelist.R
\name{read.codelist}
\alias{read.codelist}
\title{Load a Codelist (from CSV)}
\usage{
read.codelist(infile, remove.special.characters = FALSE)
}
\arguments{
\item{infile}{Path to the codelist (including name and file ending of codelist)}

\item{rm.special}{Logical; Should special characters be removed from the code column (this only looks at the columns ATC, EBM, ICD, OPS). Defaults to FALSE and gives a warning, if special characters are present.}
}
\value{
Codelist as data.frame
}
\description{
This function loads a codelist that is stored as CSV file. It assumes that the
general structure of codelists (e.g. from master codelists) of the Clinical
Epidemiology department was adhered to. The function also carries out some
basic plausibility checks and gives warnings if some common mistakes pop up.
}
